+++ python_modules/dagster/dagster/core/container_context/config.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, Optional, cast
 
 from dagster import Field, Permissive
-from dagster.config.validate import process_config
+from dagster._config.validate import process_config
 from dagster.core.errors import DagsterInvalidConfigError
 
 SHARED_CONTAINER_CONTEXT_SCHEMA = Permissive(
