+++ python_modules/dagster/dagster/core/test_utils.py
@@ -15,8 +15,8 @@ import yaml
 from dagster import ModeDefinition, Shape
 from dagster import _check as check
 from dagster import composite_solid, fs_io_manager
-from dagster._legacy import pipeline, solid
 from dagster._config import Array, Field
+from dagster._legacy import pipeline, solid
 from dagster.core.host_representation.origin import (
     ExternalPipelineOrigin,
     InProcessRepositoryLocationOrigin,
