+++ python_modules/dagster/dagster_tests/cli_tests/command_tests/test_cli_commands.py
@@ -25,11 +25,11 @@ from dagster import (
     op,
     repository,
 )
-from dagster._legacy import pipeline, solid
 from dagster._cli import ENV_PREFIX, cli
 from dagster._cli.job import job_execute_command
 from dagster._cli.pipeline import pipeline_execute_command
 from dagster._cli.run import run_delete_command, run_list_command, run_wipe_command
+from dagster._legacy import pipeline, solid
 from dagster.core.definitions.decorators.sensor_decorator import sensor
 from dagster.core.definitions.partition import PartitionedConfig, StaticPartitionsDefinition
 from dagster.core.definitions.sensor_definition import RunRequest
