+++ python_modules/dagster/dagster_tests/cli_tests/test_asset_commands.py
@@ -2,8 +2,8 @@ import pytest
 from click.testing import CliRunner
 
 from dagster import AssetKey, AssetMaterialization, Output, execute_pipeline
-from dagster._legacy import pipeline, solid
 from dagster._cli.asset import asset_wipe_command
+from dagster._legacy import pipeline, solid
 from dagster.core.instance import DagsterInstance
 from dagster.seven import json
