+++ python_modules/dagster/dagster_tests/core_tests/config_types_tests/test_config_enums.py
@@ -12,9 +12,9 @@ from dagster import (
     PipelineDefinition,
     execute_pipeline,
 )
-from dagster._legacy import pipeline, solid
 from dagster._config import Enum as ConfigEnum
 from dagster._config.validate import validate_config
+from dagster._legacy import pipeline, solid
 
 
 def define_test_enum_type():
