+++ python_modules/dagster/dagster_tests/core_tests/config_types_tests/test_config_schema.py
@@ -2,8 +2,8 @@ import pytest
 
 from dagster import Field, composite_solid
 from dagster._check import CheckError
-from dagster._legacy import solid
 from dagster._config import ConfigAnyInstance
+from dagster._legacy import solid
 
 
 def test_solid_field_backcompat():
