+++ python_modules/dagster/dagster_tests/core_tests/config_types_tests/test_type_printer.py
@@ -1,5 +1,4 @@
 from dagster import Field, Int, Map, Noneable, PipelineDefinition, ScalarUnion, String
-from dagster._legacy import solid
 from dagster._config import (
     config_schema_snapshot_from_config_type,
     get_recursive_type_keys,
@@ -7,6 +6,7 @@ from dagster._config import (
     resolve_to_config_type,
     snap_from_config_type,
 )
+from dagster._legacy import solid
 
 
 def assert_inner_types(parent_type, *dagster_types):
