+++ python_modules/dagster/dagster_tests/core_tests/snap_tests/test_pipeline_snap.py
@@ -12,8 +12,8 @@ from dagster import (
     Selector,
     Shape,
 )
-from dagster._legacy import pipeline, solid
 from dagster._config import Array, Bool, Enum, EnumValue, Float, Int, Noneable, String
+from dagster._legacy import pipeline, solid
 from dagster.core.snap import (
     DependencyStructureIndex,
     PipelineSnapshot,
