+++ python_modules/dagster/dagster_tests/general_tests/compat_tests/test_back_compat.py
@@ -15,8 +15,8 @@ import sqlalchemy as db
 from dagster import AssetKey, AssetMaterialization, Output
 from dagster import _check as check
 from dagster import execute_pipeline, file_relative_path, job
-from dagster._legacy import pipeline, solid
 from dagster._cli.debug import DebugRunPayload
+from dagster._legacy import pipeline, solid
 from dagster.core.definitions.dependency import NodeHandle
 from dagster.core.events import DagsterEvent
 from dagster.core.events.log import EventLogEntry
